app/backend/requirements.txt
fastapi==0.85.0
uvicorn==0.18.2
python-dotenv==0.21.0
groq
mysql-connector-python==8.0.30
